GFX entry panel fixes
Converting to paletted PNG from the GFX entry panel removed transparency
because when writing PNG it relied on the status of the tRNS menu item
rather than the content of the image, and converting from the GFX panel
did not update the PNG-related menu items.
Optimizing PNG from the GFX panel marked it as modified, prompting for a
save, which caused rewriting the entry and destroying any optimization
done by external tools.