Add arrows & drop-down list to texture tabs
This fixes #19. Also prevented TextureXPanels from hiding or showing the
texture menu directly, since the order in which they are hidden and
shown when changing tabs cannot be relied upon. For example, going from
Doom's TEXTURE1 to TEXTURE2 would result in the texture menu
disappearing as TEXTURE2 was shown before TEXTURE1 was hidden...