Add support for bitflags and enums in the arg editing panel.
Support was already mostly present in the configuration file format and
arg_t structure, but unused.  I used it for a couple common action
specials and updated the dialog to show custom controls when it can --
which should also get rid of the need for very long arg descriptions.

The line info overlay also now shows textual descriptions of these
argument types.

The property grid is not yet aware of any of this; it could easily be
made to support enum-style args, but there's no obvious way to cram
flags into such a small interface.