Don't traverse the same line twice in SectorBuilder::traceOutline.
This avoids some infinite loops with bad architecture.  There's already
protection against this in the form of limiting to 10,000 steps, but
that still freezes SLADE for several minutes in some cases; this ensures
we stop trying to draw the outline at all as soon as we run out of new
edges to follow.