Move activation-related props onto the Special tab.
I can't tell you how many times I forget to make a door repeatable
because the checkbox is buried in the middle of the General tab.

This also rejiggers the code a bit so these props get their own little
box.

As a free bonus, now they appear on the "trigger" line in the infobox
too!