Fix broken sectors being generated in some cases.
Consider this architecture, where the asterisks indicate the line being
drawn:

    +--------+--+
    |        |  |
    |  +--+  |  |
    |  |  |  |  |
    |  |  |*****|
    |  |  |  |  |
    |  +--+  |  |
    |        |  |
    +--------+--+

SLADE would correctly split the line into a left half and a right half,
then trace a total of three sectors.

The left half would appear in the sector's list of edges twice: once for
its front side, once for its back side.

setLineSector attempts to cleverly flip a line automatically if it has
only a back side, and no front side.  But if setLineSector is called on
the left half's back side first, this results in it having a /front/
side and no back side.  Then setLineSector is called on the front side,
but the line already has a front side, so nothing happens.  The result
is that the left half of the line is one-sided, and the sector is
broken.

Moving the flip to a loop at the end of mergeArch resolves this problem
and solves a number of cases where drawing overlapping shapes would
produce broken geometry.