Split lines by vertex before splitting on intersections.
The intersection code is not really well-suited for lines that overlap
in curious ways; splitting by vertex first ensures that any overlaps are
exact duplicate lines.

This resolves the following problem:

    +-------------+
    |             |
    +-------------+

       +--+  +--+
       |  |  |  |
       |  +--+  |
       |        |
       +--------+

Move the U-shaped piece upwards so its top edges overlap with the bottom
edge of the rectangle.

Previously, this would generate several extraneous invisible lines with
bad sector references, and also confuse the sector builder into creating
a third sector in the middle of the U, which is probably not what the
map author intended.