Implement most of the remaining ZDoom slope mechanisms.
There are still some problems here that can seemingly only be resolved
by taking ZDoom's approach and always computing every sector's slope in
one big pass.  Upon reflection, that might be a faster way to do it
anyway, since in the general case most sectors aren't sloped and we're
repeating a lot of work for no reason.

This also uses templates over an enum to cut down on code duplication
between floors and ceilings.  I'm not sure how idiomatic this is, but I
like the results.