Much much better way of recomputing map specials.
What we really want is a guarantee that map specials are always
recalculated after any kind of map edit, with minimal churn during a
bulk operation.  Lo and behold, we already have something that's called
after every major edit: the undo system!  So now we just do the work
after an undo record is created, after an undo/redo, or after the map is
initially loaded.