Don't setModified() when changing a sector's plane.
This is causing the map editor to think a map with slopes is modified as
soon as it's opened, because applying slope effects changes planes.  But
the planes aren't actually part of the map data, so it doesn't make
sense to consider changes to them as changes to the map itself.