Fix deletion of arg0str on things
The default for this property was 0, and the code that skips default
property values uses the type of the default rather than the type of the
property.  So every time a map is saved, the /integer/ value of arg0str
was compared to 0, which always came out true and caused the property to
be lost.