Fix major cause of focus stealing in map editor
This has been driving me up the goddamn wall.

When the mouse highlights a new object, and that object is a different
type from whatever was last shown in the props panel, the props panel
resets itself.  This involves calling RemovePage() on the side1 and
side2 tabs, just in case they were present.  For some asinine reason,
RemovePage() unconditionally changes the focus to one of the remaining
tabs, even if the notebook didn't have the focus in the first place!

Freezing and hiding the notebook while removing the pages seems to fix
it; cheers, some other poor soul on wxwidgets's Trac, six years ago.