Only do the directory check when the app itself is activated.  #343
Previously this happened whenever the main window was activated, which
included closing the dialog asking whether to apply changes!  There was
a workaround for this, but it inadvertently prevented any further checks
from running until the archive was saved, which could result in lost
work.

This should fix the problem cleanly, and also avoid some spurious
checks.