Fix an obscure map tracing error
If the vertex being considered happens to be directly east of one end of
a candidate line, and the other end of that line is further east, the
distance between them would be computed as negative.  That line would
therefore always become part of the sector, even if there are other much
closer lines between.

fabs() fixes this.