Expand on labels for speed arguments
It seems that Boom's generalized speeds are different for different
specials, so they need to be part of the argument, not hard-coded into
the app.

Also changed the UI for a speed argument to use a combo box, so you can
see all the labeled speeds, even if not on Windows where the slider
shows tick marks.