Allow SActions to 'reserve' a range of wx ids
This allows a single SAction to handle multiple menu/toolbar items with a range of wx ids.

SActionHandler now has a wx_id_offset protected variable which is set to the wx id 'offset' (the difference between the SAction wx id and the actual wx id that was activated) before 'handleAction' is called.