Fix custom properties errantly being identified as bools
Problem arose when doing the following:

1. In sector mode, add a string custom property like "moreids".
2. Select a sector, click the property field, and immediately click away.
3. Switch to another mode and highlight an object so the props panel refreshes.
4. Switch back to sector mode and highlight the sector you chose.

The "moreids" property will be added, but it'll be given a boolean
editor field.  The sector from step 2 has the custom property in its
properties map, but the value is left unset -- because in C++ maps,
merely trying to read a key creates a new blank value.  The default type
of a Property is boolean, so that sector had a boolean "moreids" with no
value, and when the prop grid tried to load that sector's properties, it
dutifully created a boolean field.  Examining any other sector would
leave the field unchanged, because the prop grid thinks it already
exists.

The easy fix was to just not create the field in openObjects if the
object doesn't have a value set.

This could be a little more robust -- zany things will still happen if
a map genuinely has the same property with different types, and your
"moreids" custom field won't appear again until you highlight a sector
that has it.  But the immediate problem is fixed.