Change how saving zip archives is handled
When saving a zip, previously we were opening the existing zip file from the hard drive to get pre-compressed data for unmodified entries. However if the zip was modified on the disk since it was opened in SLADE, we could end up getting the wrong data (as the indices on the zip will have changed).

Instead, when opening a zip, make a copy of it in the SLADE temp folder and use that to get pre-compressed data when saving, as it is very unlikely this file will be modified outside of SLADE.

Should fix #477 and possibly #444 too