Fix crash when opening maps with slopes in Windows
This one is very strange, it looks like the combination of the templated functions and the fpoint3_t constructor that takes an fpoin2_t and a z value were the culprit. I'm assuming this is some kind of obscure compiler bug as it doesn't seem to happen with gcc or clang.