Try not to create a sector out of an enclosed void
I regularly ran into this problem when, say, wanting to move a recessed
switch from one wall to another.  I would want to copy the switch
sector, paste it in the void, and then rotate it.  But pasting would
create a sector out of the entire void, and then deleting the new sector
would leave a large chunk of my map untextured.  Yikes.

This seems to work in simple cases, and is no worse than before in
convoluted cases.