Add 'Run Map from Here' feature(s)
New 'Run Map from Here' action on the 2d editing modes context menu, which will move the player start to the mouse cursor position when running the map.

When 'Run Map' is launched via the map editor, a new checkbox shows which when ticked will move the player start to the current 3d mode camera position (and direction). This option is saved between sessions via a cvar.

Also made the 'Run' button focused by default on the run dialog.

Closes #488