Fix spurious segfaults with camera gravity
The camera gravity is updated on a timer, so it could be called before
the sector flats were calculated.  The easy fix was to use the
original extra_floors instead of the sector flats, which turned out to
be easier anyway.

As an added bonus, now gravity only checks for solid 3D floors.