Fix sector drawing bug in #542 (#548)
* Restore LOG_DEBUG in debug builds

Seems DEBUG isn't actually defined in debug builds?  Not sure why I
thought it was.

* Fix an edge case in sector building.  Fixes #542

After tracing an outer edge, SectorBuilder fires out a ray from the last
vertex to try to find a corresponding inner edge.  Tracing continues
from the first line that ray hits.

If the ray hits a vertex shared by two lines, it was possible to
continue tracing from the further of those two lines, which could have
been the _inner_ edge of some other shape.

Adding a tiebreaker, which checks the actual distance to each line,
fixes this.