Add build configuration for WinXP
This one removes cURL completely (and wxEmail along with it) as I can't get it to compile with WinXP compatibility.

The NOCURL preprocessor define will remove all cURL stuff and change the crash report dialog to remove the email sending stuff.

This is probably the last thing I'm going to do for WinXP compatibility, if something breaks it in future I won't be fixing it (myself, at least) and official builds will no longer support it.

Also, openal32.dll is actually still needed, apparently I have it in my windows/system32 folder >_>