Begin separating some UI stuff
Time to start on a long overdue source cleanup, to avoid having to #include compile-time-heavy wx stuff as much as possible, among other things. Currently the wx UI is much too intertwined with the backend resulting in a bunch of heavy #includes that shouldn't really be needed.

It'd also be nice to get everything to a state where I could rip out and replace the current wx UI (not that it would happen :P) if I wanted without messing up the back end too much.

To start:
- Added new UI namespace for functions invoking the general UI, beginning with the SplashWindow
- Split out some non-UI-related functions from MainWindow to a new MainEditor namespace