Clean up some parser/tree stuff
Added a few helper functions to the Parser class to avoid the need to cast from STreeNode to ParseTreeNode so much. Eventually I'd like to implement a templated tree class or something to avoid this, but that gets kinda complicated and the current one works well enough, albeit messily.

I've also adopted a new header/boilerplate comment style which I'll use from now on. I know it's not really necessary but I was getting sick of having visual studio screw with the formatting when copy+pasting them. As with the new class member naming I'll update files with the new style as I go along rather than do everything at once, which would take forever :P