Game Config Code Cleanup (#726)
* Begin cleaning up / improving game config stuff

Starting with the action special and thing type args stuff

* Replace GameConfiguration singleton implementation

The class is now just called 'Configuration' within the 'Game' namespace. Access via 'Game::configuration().' rather than 'theGameConfiguration->' macro

* Consolidate 'feature support' stuff in Configuration class

Now we have enums/maps for general and udmf features rather than a bunch of bool variables and getters

* Fix crash when opening line properties

Need to be careful when changing arguments/return types from pointers to references

* Clean up some parser/tree stuff

Added a few helper functions to the Parser class to avoid the need to cast from STreeNode to ParseTreeNode so much. Eventually I'd like to implement a templated tree class or something to avoid this, but that gets kinda complicated and the current one works well enough, albeit messily.

I've also adopted a new header/boilerplate comment style which I'll use from now on. I know it's not really necessary but I was getting sick of having visual studio screw with the formatting when copy+pasting them. As with the new class member naming I'll update files with the new style as I go along rather than do everything at once, which would take forever :P

* Move basic game/port definition stuff out of Configuration class

Now these are just handled/stored in the Game namespace

* Cleanup/Improve ActionSpecial handling

No more need to as_t in Configuration, just keep the number etc in the ActionSpecial itself. Also changed the TagType stuff to an enum class, will need to convert ThingType over to it next (and properly implement the thing pathing which is currently a bit of a hack)

* Cleanup/Improve ThingType handling (WIP)

Similarly to the previous commit, this removes the need for the container struct tt_t in Game::Configuration, among (many) other things.

WIP commit as some things are currently not (re-)implemented:
- DECORATE parsing
- Pathed things

* Fix wrong port being selected in MapEditorConfigDialog

* Move Decorate handling stuff out to a separate source file

With this DECORATE parsing should be working again

* (Re-)implement missing thing path tagging stuff

It's now a bit more hard-coded but they are very much special cases anyway.

* Compilation fixes

* Refactor/Cleanup UDMFProperty stuff

* Fix 3d mode thing sprites not displaying

* Finish 'Game' stuff refactoring/cleanup (for now)

Probably could clean up/modernise a bit more but it'll do for now