Export more things for scripting
Can now manupulate the selection in the map editor, among other things. Also removed the ItemSelection type since it's pretty useless on its own for scripting purposes, the functions make more sense being part of the MapEditor.

Documentation also updated with the new stuff and improved the theme a bit