Update scripting documentation
Changed the docs theme to 'mkdocs-material' (with some tweaks via slade-material.css). The scrolling navigation menu of the previous theme was kind of annoying. This one can be contracted/expanded as necessary and shows the page table of contents on the right rather than adding them to the navigation with all the rest.