Rework scripting naming scheme
Decided to go with something closer to the SLADE sources, so namespaces, types and enums are in PascalCase, function names and properties are in camelCase, and parameters and variables use underscores. Properties are kind of the odd one out as c++ obviously doesn't have them, I went with camelCase for them since they are kind of on the same level as functions for the most part (public members of a type)

Also renamed the slade namespace to App, since Slade just looks wrong (:P) and typing SLADE isn't fun.