Implement Entry Scripts feature
Similar to archive scripts, an entry script is run on the selected entries in the current archive tab. Entry scripts take an array of entries rather than a single one like archive scripts do.

Entry scripts can be run from the entry list context menu, or the 'Entry' menu on the main window