Start on a general UI cleanup and better High-DPI support (again)
Since the high dpi branch is pretty old, and I wasn't entirely happy with having the SM() macro sprinkled everywhere I've gone with a bit of a different approach this time.

In the UI namespace there is now a bunch of pixel scaling and 'standard' size stuff to both support high DPI and standardise the UI layout spacing a bit.

I've also added a new WxUtils namespace (replacing the old WxStuff.h) with some functions to help with common layout needs like putting labels next to controls and laying out a bunch of widgets vertically/horizontally with standard spacing.

In this initial commit I've gone through and updated all the stuff in the src/Dialogs folder, and also reworked the preferences panel a little bit (eventually I'd like to simplify it a bit more and replace the tree with something a bit nicer)