Improved high DPI scaling in Windows and general UI cleanup (#810)
There are probably still a few things that don't scale quite right but for the most part things should be much improved in Windows with a DPI higher than 100%.

This also cleans up and refactors a bunch of source files and adds some general UI utilities to help keep things more consistent such as spacing etc.



** Individual commit history: **

* Start on a general UI cleanup and better High-DPI support (again)

Since the high dpi branch is pretty old, and I wasn't entirely happy with having the SM() macro sprinkled everywhere I've gone with a bit of a different approach this time.

In the UI namespace there is now a bunch of pixel scaling and 'standard' size stuff to both support high DPI and standardise the UI layout spacing a bit.

I've also added a new WxUtils namespace (replacing the old WxStuff.h) with some functions to help with common layout needs like putting labels next to controls and laying out a bunch of widgets vertically/horizontally with standard spacing.

In this initial commit I've gone through and updated all the stuff in the src/Dialogs folder, and also reworked the preferences panel a little bit (eventually I'd like to simplify it a bit more and replace the tree with something a bit nicer)

* Scale SToolBar/Buttons with DPI

* Move generic UI controls to UI/Controls source folder

* Fix 3d mode mouselook with high dpi in windows/linux

* Scale main window, entry panels and texture editor with DPI

Also the usual cleanup/refactoring/etc, and fixed some minor issues

* Fix zoom control in the texture editor

* More DPI scaling and cleanup

Text editor, misc. UI controls, entry list

* Add large versions for all icons

Most are just lazy rescales using hq2x or xbr2 depending on the icon. Eventually i'd like to remake all the icons as svg but that's a pretty big undertaking

* Add close button to find+replace panel

resolves #805

* Scale tabs and docked window borders with DPI

* Scale list icons with DPI

* Scale map editor (wx) UI with DPI

* Scale opengl fonts with DPI

* Scale splash window with DPI

* Scale script manager window with DPI

* Various minor DPI scaling improvements