Fix UV coordinates and scale of scaled textures
Fix UV coordinates and scale for quads that use scaled textures.
Set sx and sy to the scale specified in the texture definition.
Add world_panning field to GLTexture so that the "World panning" flag is passed to the renderer, and set it in MapTextureManager.getTexture/getFlat.
Use GLTexture worldPanning to determine how UVs of quads should be manipulated.