Fix build under GCC
TextureXEditor.h failed to build because unique_ptr<T> can't be used
with incomplete types...  or, it can, sometimes, and this seems like one
of those times, but it didn't work anyway.  Including the header so the
type (UndoManager) is fully defined is an easy fix.

Some code was passing a vector directly to EntryOperations functions
that accept a vector by reference, and GCC objected.  Apparently the
problem is just that a non-const value will presumably be modified,
which makes no sense when the value is a temporary, so making the
argument a const reference fixes it.