Fix a crash when opening the map editor
SImage::create resizes its buffer and uses bpp() to determine the new
size, but type_ is reassigned AFTER bpp() is called, so the old value is
still used.  This caused a write beyond the end of the buffer when
loading grayscale + alpha PNGs (in my case, thing/flag.png), which then
caused malloc to abort a short time later.