Fix Linux build and a map editor crash (#995)
* Fix build under GCC

TextureXEditor.h failed to build because unique_ptr<T> can't be used
with incomplete types...  or, it can, sometimes, and this seems like one
of those times, but it didn't work anyway.  Including the header so the
type (UndoManager) is fully defined is an easy fix.

Some code was passing a vector directly to EntryOperations functions
that accept a vector by reference, and GCC objected.  Apparently the
problem is just that a non-const value will presumably be modified,
which makes no sense when the value is a temporary, so making the
argument a const reference fixes it.

* Fix a crash when opening the map editor

SImage::create resizes its buffer and uses bpp() to determine the new
size, but type_ is reassigned AFTER bpp() is called, so the old value is
still used.  This caused a write beyond the end of the buffer when
loading grayscale + alpha PNGs (in my case, thing/flag.png), which then
caused malloc to abort a short time later.