Move MSVC files out of 'build' directory
Now it's all in just 'msvc'. This allows the 'build' dir to be added to .gitignore completely, since many CMake-based IDEs will use build as a default intermediate dir, and I was getting sick of it spamming up git whenever I used them.

The MSVC project will also now use build as an intermediate directory, instead of the dir that the vcxproj is in.

I've also just removed the old codelite project since it hasn't been used or updated in many years.