Expand "Patches" subtree in image browser dialog opened from a TEXTURES editor
When editing a TEXTURES lump, adding a patch opens an image browser, only showing images in the Patches namespace. Because the tree panel on the left is fully collapsed, it is not obvious that you can also use sprites, flats, and other images outside the Patches namespace (I discovered this by accident). This minor UI change should make it more obvious.