Fix applying floor/ceiling plane UDMF properties to triangular sectors
If a sector was triangular, the sector plane would be overridden by the zfloor/zceiling properties, or the sector flat height if no zfloor/zceiling property values were specified.
Also, be more lenient in applying UDMF sector flat planes if any sector is missing the A, B, C, or D value.