Ignore UDMF floor/ceiling plane properties if they are all default
Since editing a sector's properties will add all UDMF props to the sector, this caused any edited sector to display wrong in 3d mode. Now the values need to be non-default for it to preview the slope. This means that if a sector legitimately needs a flat plane at height 0 it won't work correctly, but it's better than the alternative.

I think in 3.2.0 I'll change the way UDMF properties are implemented, the way default values are currently handled is a mess