Namespace cleanup/rework
Reworked namespaces to be all lowercase, and have everything SLADE-specific be within a base 'slade' namespace.

This is more in-line with 'standard' c++ naming, and also allows for some extra options eg. Classes and namespaces with the same name. Also helps visually differentiate between calling a Class' static function and a namespace function