MapEditor: fix 2D -> 3D line selection conversion
When determining which parts of the line to select in 3D mode, we check what
parts of the line should be visible and add those to the selection.  Prior
to this change, we'd always end up including the front middle and back
middle, because front->texMiddle() is never empty.

Switch the empty checks to comparisons against TEX_NONE instead, so that the
invisible middle parts don't get selected.  This fixes deselecting of doors
and "windows" or steps in 3D mode such that the lines are properly
deselected once all visible parts are deselected.