Small collection of fixes (#1297)
* Log: flush the log on write

Avoid losing log messages if the application crashes before the buffer gets
flushed next.

* MapEditor: avoid resetting special args on applyValue()

args are no longer stored in the object's hash map; they now have their own
std::array.  As such, hasProp() no longer works on them, and we don't need
to do this anyways: args are already zero-initialized.

Fixes #1293

* OpenGL: check errors from glewInit()

This avoids hiding problems that later result in crashes.  One example of
such an issue is the default change in wxWidgets 3.1 to use EGL by default,
which breaks if glew isn't also compiled to use EGL.

* MapEditor: fix 2D -> 3D line selection conversion

When determining which parts of the line to select in 3D mode, we check what
parts of the line should be visible and add those to the selection.  Prior
to this change, we'd always end up including the front middle and back
middle, because front->texMiddle() is never empty.

Switch the empty checks to comparisons against TEX_NONE instead, so that the
invisible middle parts don't get selected.  This fixes deselecting of doors
and "windows" or steps in 3D mode such that the lines are properly
deselected once all visible parts are deselected.

Co-authored-by: Kyle Evans <self@kevans.dev>