ArchivePanel: deleted windows should be replaced in the splitter (#1296)
The sizer contains the splitter, which contains the two panes.  When an
archive entry is deleted, the splitter's the one that we need to be
replacing it in.

This fixes an assertion when an archive entry is deleted, then another one
is selected. This regression seems to be from eb6eeb5db6fa, when the
splitter was introduced.

Co-authored-by: Kyle Evans <self@kevans.dev>