Remove 'email stack trace' function from crash dialog
This feature hasn't really worked out as well as I'd like, 99% of the time the stack traces received are blank or useless without context, so I think in the time it's been implemented I've fixed something like 2 crashes from them :P

Instead of the 'send crash report' button I've added a button to open GitHub and create an issue. I've also changed the dialog to be created on startup and shown after a crash, since creating it *after* a crash happens could be problematic in some cases.

This change means curl is no longer required to build in windows, which is a plus because it's huge if you include openssl too.