Add cvar to allow duplicate names in zip archives (not saving)
zip_allow_duplicate_names - defaults false

Not going to add it to the preferences dialog but if anyone *really* wants to be able to do it for editing purposes then they can enable it via the config/console.

It will still disallow saving zips with duplicates.

Also includes some minor cleanup