Update for wx3.1.6+ compatibility
Whew, what a mess.

wx3.1.6 introduces wxBitmapBundle and a bunch of its own DPI handling, which completely broke a bunch of SLADE's custom handling.

Luckily it mostly seems limited to icon/image stuff, so changing getIcon to return a wxBitmapBundle and updating code to work with that seems to work for most things from what I can tell.

Still a few things need to be done:
- Test it still compiles for pre-3.1.6 wx
- Test in linux
- (Re)implement disabled colouring for tab/dock close buttons etc.