Make all ArchiveEntry data access const
We don't want an entry's data changing without going through the correct functions (eg. ArchiveEntry::importMem) to update the entry status, send signals, etc.

As always, adding const to something that is used a lot has a cascading effect on everything else, so this commit has a *lot* of changes to add const where appropriate, and work around the few places an entry's data was accessed+modified directly