Fix crash when reading copyright ID3v2 on wxWidgets UTF-8 build
When reading the copyright ID3v2 of a music file, in this piece of code:
    wxString::Format("Copyright \x00A9 %s\n", copyright);
The 1st argument is converted to a wxString using the current locale encoding.

If the encoding is UTF-8 (typically Linux), this string is invalidly encoded
and ends up causing a crash inside wxString::format.

Explicitly specify the encoding to avoid this.