Rework entry type config loading
Instead of just loading each file in the entry_types res dir, read entry_types.cfg which #includes them. This allows us to load them in a specific order as some are based on others that may not necessarily be in the same file.

Doesn't affect loading of user-defined entry types (if anyone even uses that feature which I doubt)