Remove webview start and docs pages
This is a bit of an experiment to remove the need for wxWebView (and wxHtml), since the start page doesn't really need to be a full blown web page and I'm not sure anyone ever used the embedded docs page (not like it's hard to just have them open in a browser).

This replaces the start page with one using just wxWidgets controls. It doesn't have the latest news post and tip of the day like the previous one but should have the most important things (might re-add tips later as a separate window or something).

I also want to switch the update check back to use wxWidgets for the web call instead of SFML, since the SFML one seems to cause crashes on linux sometimes