Show the splash window atop the main window where possible
At least on GTK, this avoids giving the splash window its own (untitled,
icon-less) taskbar entry, and also prevents clicking on the main window
and accidentally raising it above the splash window.  It does mean the
splash window needs to be recreated after app init, but it's extremely
cheap so that should be fine.