Avoid adding to the configuration when looking up a bogus UDMF prop
The properties side panel looks up all the known UDMF properties on,
say, the current line.  This includes faux side properties like
side2.sector.  If a line actually has a back side, it'll defer that
lookup to it; otherwise it'll treat 'side2.sector' like a real property
name and fall back to MapObject::intProperty.

That calls Configuration::getUDMFProperty, which looks the name up in a
map.  But looking up a missing key in a std::map creates a new value
with the default constructor.

As a result, mousing over a one-sided line would add several dozen blank
UDMFPropertys to the configuration, all named after prefixed side
properties.  The visual effect was a huge stack of blank categories in
the sidebar, manifesting as a sea of medium gray.  Apparently
wxPropertyGrid gets a bit confused with empty names and fails to return
the existing group each time.

Anyway, the fundamental solution is for getUDMFProperty to never create
new canonical properties in the first place.  It can already return
nullptr, and every caller appears to check for it anyway, so this should
be a safe change.