Fix magic DECORATE/ZScript comments sometimes eating too many tokens
I recently ported a bunch of DECORATE objects to ZScript, and their
sprites stopped appearing in SLADE.  Some investigation revealed that
the problem was that I used "//$NotAngled" as the last magic comment in
many of them, and because the tokenizer tries to handle this itself, it
would dutifully always consume the next token -- which was "states" on
the following line.

I couldn't figure out a way to fix this within the tokenizer, since it
fundamentally can't know what it's supposed to be looking for.  Even if
it did know the syntax of every magic comment, the same problem would
happen if someone put "//$Sprite" on its own line before a states
block.  That's incorrect, of course, but it's within a comment, so it
shouldn't affect parsing of the rest of the file.

So I've changed it to do a little sub-parse of the comment instead,
and shared that code with the DECORATE parser as well.

I ended up not using strutil::split(..., true), but there's a free
feature, I guess.