Add arg types for tics and octics
Inspired by realizing that this whole time I've been convinced the
standard Doom door delay is 50 tics.  It's 105.  Wait, no, that's also
wrong, it's 150.

Also found out why the combo box was weirdly tall on GTK: it's because
the slider is weirdly tall, and the combo box is stretching to match it.