Fix an infinite loop in SLADEMap::mergeArch
I had some architecture with vertices that were slightly off integer
coordinates -- I think perhaps some float cruft crept in after using
Edit objects? -- and simply moving it caused the "split lines that moved
onto vertices" loop to run forever.

You can replicate this by drawing two boxes (in a UDMF map), manually
editing one corner of one box (call it A) to have a coordinate 0.05 off
an integer, and then dragging a corner of the second box (call it B)
onto it.  You'll get stuck as follows:

1. One of the affected edges of the second box (BC) will be detected as
   being within 0.1 units of A, and so that line will be split.

2. The split will create two lines: AB and AC.  AB is only 0.05 units
   long, and AC is nearly identical to the original line.

3. These new lines will be added to `connected_lines` so the loop can
   examine them for any further vertex collisions.

4. On a later iteration, AC will be detected as being within 0.1 units
   of B, and so it will be split -- into a duplicate AB and BC.  These
   will also be added to `connected_lines`, and so on.

The simplest fix seems to be to make `mergeVerticesPoint` use the same
fuzz radius as the line-splitting, so the two vertices are collapsed
before any of this can happen.

That said, float precision makes it possible that this could still
happen under extremely specific circumstances.  A more robust fix might
be to identify duplicate lines upfront and also as they're created, but
that's a bit more complicated.