Add rough support for deleting items from the prop grid
By "delete" I mean "reset to default values", since there's no real
notion of a missing property.  But it's handy for, say, nuking the
special + args off of a line.

This doesn't delete custom properties, either, because that breaks
things; currently the list of custom properties persists until the whole
grid is rebuilt.