Allow filtering action specials by name
This is pretty jank in a lot of ways, but the old version was, too, and
I don't think this is any worse.  The keyboard navigation in particular
leaves a lot to be desired.  (Simply pressing up/down in the tree
doesn't even work right?  That's definitely not my fault.)

Anyway, now you can do all sorts of things like filter the vanilla
special list down to just S1, or find the locked doors by typing "red",
or find the ZDoom portal specials even though you could make a good
argument for finding them in Teleport or Rendering or even Sector.