Fix "Change Special" dialog being empty (on GTK?)
I don't know why, but something about the order of events here produced
a GTK warning and prevented the inner panels from being added to the tab
control, so the dialog was just two big blank tabs.

It shouldn't matter what the initial parent is since we change it
immediately, and parenting to the dialog itself works fine.