Support wxWidgets 3.3 and dark theme in Windows (#1743)
Added the option to download/build wx via cmake using FetchContent instead of using the system wx lib (windows-only currently), so we can build with wxWidgets 3.3.0 if needed.

With wx 3.3 we can now enable the 'experimental' dark theme support in Windows. Added an option to select the theme (or go with the system preference, which is the default) and made some tweaks to support the dark theme. There are a few minor issues but it works well enough for the most part.