Switch lua binding library from sol2 to luabridge
Since sol2 appears to be pretty much abandoned by it's author and is beginning to cause some issues (not to mention the ridiculous compile times), I'm having a look at switching over to a different lua binding library, luabridge.

On the plus side, it seems to be a lot simpler than sol2 and compile times are *much* lower, it was also pretty simple to convert the code over, which for the most part was just changing function names (eg. set_function -> addFunction).

There are a couple of downsides though:
- The most significant issue is that luabridge doesn't support multiple return values, which the SLADE scripting API makes use of in a few places, so existing scripts would need to be changed to expect a table instead of the multiple values. It's not too difficult to work around since we can just use table.unpack(), but annoying all the same as it will break existing scripts.
- Luabridge uses a different method of creating new objects on the lua side, instead of Type.new() it's now just Type(). Not a huge deal but again it would break old scripts.
- It also differentiates between const and non-const objects, which could be a source of confusion if not dealt with correctly - if a function returns a const object to lua, only the const methods are available, so trying to call a non-const method on the object results in an error (and not a very helpful one either). For this I'll need to be extra careful and make sure we only deal with non-const objects in scripts, even if it means const_casting them
- Luabridge doesn't support unique_ptrs in vectors, not a huge issue as I'll just have to convert to raw pointers

So far it seems to mostly work ok, I need to go through and write some test scripts for everything to make sure it all works (and it would be good to have them anyway to ensure things continue working)

Also really need to deal with the terrible error messages from lua, clean them up a bit, and the console window is really bad at dealing with multi-line messages. Wonder if it might be an idea to change it to a condensed list view instead of a text control...